Current status. Mostly working.
