+++ .gitignore
@@ -1,5 +1,1 @@
-# PHPStorm settings
-.idea
-img/*
-!img/unknown.jpg
 node_modules/
