+++ static/.gitignore
@@ -1,0 +1,3 @@
+img/*
+!img/unknown.jpg
+!img/favicon.png
