+++ static/index.html
@@ -21,9 +21,13 @@
 
 </head>
 <body class="dark">
+    <header-comp></header-comp>
 
-<single-series id="seriesContent" class="container-fluid">
-<!--?php
+    <all-series id="seriesContent" class="container-fluid"></all-series>
+
+
+    <dialog-overlay id="overlayContent"></dialog-overlay>
+        <!--?php
 require_once 'server/Helper.php';
 require_once 'server/FileHandler.php';
 require_once 'server/DbConnectionMySql.php';
@@ -60,40 +64,41 @@ foreach ($serien as $title => $serie) {
     array_push($titelList, Helper::replaceHyphen($title));
 }
 ?-->
-</single-series> <!-- /container -->
 
 
-<script type="module" src="js/comp.js"></script>
 
 <!-- JQuery https://jquery.com/ -->
 <script src="https://code.jquery.com/jquery-3.3.1.min.js" integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8=" crossorigin="anonymous"></script>
 <!-- Bootstrap https://v4-alpha.getbootstrap.com/ -->
 <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
 <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
-<!-- Lazy Loading http://jquery.eisbehr.de/lazy/ -->
-<script type="text/javascript" src="libs/jquery-lazy/jquery.lazy.min.js"></script>
-<script type="text/javascript">
+    <!-- Lazy Loading http://jquery.eisbehr.de/lazy/ -->
+    <!-- <script type="text/javascript" src="libs/jquery-lazy/jquery.lazy.min.js"></script> -->
+    <script type="text/javascript">
    /* let allSeries = {
-    <?php
-    $first = true;
-    foreach ($serien as $serie) {
-        if ($first) {
-            $first = false;
-        } else {
-            echo ",";
+       <?php
+       $first = true;
+       foreach ($serien as $serie) {
+           if ($first) {
+               $first = false;
+            } else {
+                echo ",";
+            }
+            echo "'" . Helper::replaceHyphen($serie->title) . "': { status: '" . $serie->status . "', image: '" . $serie->image . "'}\n";
         }
-        echo "'" . Helper::replaceHyphen($serie->title) . "': { status: '" . $serie->status . "', image: '" . $serie->image . "'}\n";
-    }
-    ?>
+        ?>
     };*/
-</script>
+    </script>
 <!-- MD5 https://github.com/blueimp/JavaScript-MD5 -->
-<script type="text/javascript" src="js/md5.js"></script>
+<!-- <script type="text/javascript" src="js/md5.js"></script> -->
 <!-- My JS -->
-<script type="text/javascript" src="js/series.js"></script>
-<script type="text/javascript" src="js/cookies.js"></script>
-<script type="text/javascript" src="js/fuzzy_search.js"></script>
+<!-- <script type="text/javascript" src="js/series.js"></script>
+    <script type="text/javascript" src="js/cookies.js"></script> -->
+    <!-- <script type="text/javascript" src="js/fuzzy_search.js"></script> -->
 <script type="text/javascript" src="js/interface.js"></script>
 <script type="text/javascript" src="js/tooltip.js"></script>
+<script type="module" src="js/headerComp.js"></script>
+<script type="module" src="js/dialogComp.js"></script>
+<script type="module" src="js/seriesComp.js"></script>
 </body>
 </html>
